+++ HISTORY.md
@@ -1,5 +1,16 @@
 # Release History
 
+### 0.2.8rc1 (2024-08-07)
+----------------------------
+**Implementations**
+- add version specifiers for aliyun-log-python-sdk  
+- print more detailed error message when passing invalid parameters
+
+**Fixing**
+- get_logs using post method return the same data as before
+- fix configParser on python 3.12
+
+
 ### 0.2.0 (2021-10-17)
 ----------------------------
 **Implementations**